double to long in getter

+++ src/main/java/com/example/hotels/model/Hotel.java
@@ -88,7 +88,7 @@ public class Hotel {
         this.workingStatus = workingStatus;
     }
 
-    public double getDailyCost() {
+    public long getDailyCost() {
         return dailyCost;
     }
